Reorganizada estrutura do projeto (MVC)
