+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/model/CadastroColaborador.java
@@ -1,4 +1,4 @@
-package br.com.proway.senior.fechamentoFolha;
+package br.com.proway.senior.model;
 
 import java.util.ArrayList;
