+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/model/Pessoa.java
@@ -1,4 +1,4 @@
-package br.com.proway.senior.fechamentoFolha;
+package br.com.proway.senior.model;
 
 public abstract class Pessoa { // Abstract
